fix #167 prefetch after union
